CropRecommendationSystem/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
joblib
streamlit
flask
flask-cors
google-generativeai
python-dotenv